% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHurdleQap.R
\name{plotHurdleQap}
\alias{plotHurdleQap}
\title{plotHurdleQAP}
\usage{
plotHurdleQap(
  results,
  method = "parametric",
  rug = FALSE,
  plotTitles = NULL,
  xLabels = NULL,
  display = "terms",
  legend = TRUE
)
}
\arguments{
\item{results}{object of class HurdleQap}

\item{method}{"parametric" if results of the parametric model should be plotted; "non-parametric" for the non-parametric model}

\item{rug}{}

\item{plotTitles}{list of characters which represent the single plot titles. Dimension should be equal to two times the number of covariates.}

\item{xLabels}{list of characters which represent x-axis labels. Dimension should be equal to two times the number of covariates.}

\item{display}{only meaningful for method = "non-parametric". If display = "terms", y-axis of plot is on original (non-transformed) scale.
If display = "response", y-axis is displayed on the predicted response scale.}
}
\description{
Plot function for object from class HurdleQap.
}
\details{
plotHurdleQAP

See also function \link[hurdleqap]{hurdleQap}.
}
